homelab-builder/requirements.txt
# Homelab Builder Requirements

# Core dependencies
rich>=13.0.0
questionary>=2.0.0
pyyaml>=6.0
requests>=2.28.0

# System information
psutil>=5.9.0
distro>=1.8.0
netifaces>=0.11.0

# Optional dependencies for enhanced features
qrcode[pil]>=7.4.0  # For WireGuard QR codes
python-dotenv>=1.0.0  # For environment file handling

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
